requirements.txt
django
django-tastypie
PIL

